rename and change filenames to kebab-case
