+++ packages/repco-frontend/app/components/primitives/pager.tsx
@@ -1,5 +1,5 @@
 import type { PageInfo } from '~/graphql/types.js'
-import { NextButton, PrevButton } from './primitives/Button'
+import { NextButton, PrevButton } from '../primitives/button'
 
 export function Pager({
   pageInfo,
