+++ packages/repco-frontend/app/routes/__layout.tsx
@@ -1,10 +1,10 @@
-import Player from '~/components/player/Player'
+import Player from '~/components/player/player'
 import type { LoaderArgs } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
-import { NavBar } from '@ui/bars/NavBar'
-import { Logo } from '@ui/primitives/logo'
 import { Outlet } from 'react-router-dom'
-import { QueueView } from '~/components/player/Queue'
+import { NavBar } from '~/components/navigation/nav-bar'
+import { QueueView } from '~/components/player/queue-view'
+import { Logo } from '~/components/primitives/logo'
 import { GitHubLoginButton } from '~/routes/__layout/login'
 import { authenticator } from '~/services/auth.server'
 import { LogoutButton } from './logout'
