+++ packages/repco-frontend/app/routes/__layout/playlists.tsx
@@ -1,9 +1,9 @@
 import { ListBulletIcon } from '@radix-ui/react-icons'
 import { Outlet } from '@remix-run/react'
 import { useState } from 'react'
-import { usePlaylists } from '~/components/player/usePlaylists'
-import { Button } from '~/components/ui/primitives/Button'
-import { InputWithIcon } from '~/components/ui/primitives/Input'
+import { usePlaylists } from '~/components/player/use-playlists'
+import { Button } from '~/components/primitives/button'
+import { InputWithIcon } from '~/components/primitives/input'
 
 export default function ItemsMenuWrapper() {
   const [name, setName] = useState<string>()
