+++ packages/repco-frontend/app/routes/logout.tsx
@@ -1,7 +1,7 @@
 import { LockOpen2Icon } from '@radix-ui/react-icons'
 import type { ActionArgs } from '@remix-run/node'
 import { Form } from '@remix-run/react'
-import { IconButton } from '~/components/ui/primitives/Button'
+import { IconButton } from '~/components/primitives/button'
 import { authenticator } from '~/services/auth.server'
 
 export async function action({ request }: ActionArgs) {
